requirements.txt
google-cloud-bigquery
google-generativeai
pandas
streamlit
python-dotenv
scipy
db-dtypes
